Split helper methods by command and query
